organized the pricing code a bit
