+++ backend/schemas/billing.py
@@ -1,0 +1,8 @@
+
+from pydantic import BaseModel
+
+
+class CheckoutSessionRequest(BaseModel):
+    price_id: str
+    success_url: str
+    cancel_url: str
